Refactor helper.blogDir
